Removed redundant photometry tests
